+++ internal/platform/mmap_test.go
@@ -1,8 +1,9 @@
-package compiler
+package platform
 
 import (
 	"crypto/rand"
 	"io"
+	"runtime"
 	"testing"
 
 	"github.com/tetratelabs/wazero/internal/testing/require"
@@ -12,7 +13,8 @@ var testCode, _ = io.ReadAll(io.LimitReader(rand.Reader, 8*1024))
 
 func Test_mmapCodeSegment(t *testing.T) {
 	requireSupportedOSArch(t)
-	newCode, err := mmapCodeSegment(testCode)
+
+	newCode, err := MmapCodeSegment(testCode)
 	require.NoError(t, err)
 	// Verify that the mmap is the same as the original.
 	require.Equal(t, testCode, newCode)
@@ -20,9 +22,9 @@ func Test_mmapCodeSegment(t *testing.T) {
 
 	t.Run("panic on zero length", func(t *testing.T) {
 		captured := require.CapturePanic(func() {
-			_, _ = mmapCodeSegment(make([]byte, 0))
+			_, _ = MmapCodeSegment(make([]byte, 0))
 		})
-		require.EqualError(t, captured, "BUG: mmapCodeSegment with zero length")
+		require.EqualError(t, captured, "BUG: MmapCodeSegment with zero length")
 	})
 }
 
@@ -30,19 +32,26 @@ func Test_munmapCodeSegment(t *testing.T) {
 	requireSupportedOSArch(t)
 
 	// Errors if never mapped
-	require.Error(t, munmapCodeSegment(testCode))
+	require.Error(t, MunmapCodeSegment(testCode))
 
-	newCode, err := mmapCodeSegment(testCode)
+	newCode, err := MmapCodeSegment(testCode)
 	require.NoError(t, err)
 	// First munmap should succeed.
-	require.NoError(t, munmapCodeSegment(newCode))
+	require.NoError(t, MunmapCodeSegment(newCode))
 	// Double munmap should fail.
-	require.Error(t, munmapCodeSegment(newCode))
+	require.Error(t, MunmapCodeSegment(newCode))
 
 	t.Run("panic on zero length", func(t *testing.T) {
 		captured := require.CapturePanic(func() {
-			_ = munmapCodeSegment(make([]byte, 0))
+			_ = MunmapCodeSegment(make([]byte, 0))
 		})
-		require.EqualError(t, captured, "BUG: munmapCodeSegment with zero length")
+		require.EqualError(t, captured, "BUG: MunmapCodeSegment with zero length")
 	})
+}
+
+// requireSupportedOSArch is duplicated also in the compiler package to ensure no cyclic dependency.
+func requireSupportedOSArch(t *testing.T) {
+	if runtime.GOARCH != "amd64" && runtime.GOARCH != "arm64" {
+		t.Skip()
+	}
 }
